feat: move mock to internal + try fix linter on CI
